Move actionselection tests to thalamus and bg test
