+++ nengo_spa/modules/tests/test_thalamus.py
@@ -7,6 +7,30 @@ from nengo_spa.exceptions import SpaNetworkError
 import numpy as np
 
 
+def test_thalamus_basic(Simulator, plt, seed):
+    with nengo.Network(seed=seed) as net:
+        bg = spa.BasalGanglia(action_count=5)
+        input = nengo.Node([0.8, 0.4, 0.4, 0.4, 0.4], label="input")
+        nengo.Connection(input, bg.input, synapse=None)
+
+        thal = spa.Thalamus(action_count=5)
+        nengo.Connection(bg.output, thal.input)
+
+        p = nengo.Probe(thal.output, synapse=0.01)
+
+    with Simulator(net) as sim:
+        sim.run(0.2)
+
+    t = sim.trange()
+    output = np.mean(sim.data[p][t > 0.1], axis=0)
+
+    plt.plot(t, sim.data[p])
+    plt.ylabel("Output")
+
+    assert output[0] > 0.8
+    assert np.all(output[1:] < 0.01)
+
+
 @pytest.mark.slow
 def test_thalamus(Simulator, plt, seed):
     model = spa.Network(seed=seed)
